+++ config_spec/scout_config.ini
@@ -9,14 +9,14 @@ collaborators = force_list(default=[])
 madeline = string(default=None)
 
 #[gene_lists]
-#  
-#  [[__many__]]
-#    name = string()
-#    version = float()
-#    full_name = string()
-#    date = string()
 #
-#[individuals]
+# [[__many__]]
+#   name = string()
+#   version = float()
+#   full_name = string()
+#   date = string()
+#
+##[individuals]
 #  
 #  [[__many__]]
 #    name = string()
