+++ scout/commands/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function, absolute_import, unicode_literals
 from scout.commands.load import load_mongo
-from scout.commands.wipe_and_load import wipe_and_load
+from scout.commands.wipe_and_load_mongo import wipe_and_load
 from scout.commands.wipe import wipe_mongo
