+++ scout/commands/wipe_and_load_mongo.py
