+++ tests/vcf_examples/1/scout_config_test_clinical.ini
@@ -8,6 +8,7 @@ owner = CMMS
 collaborators = PID
 madeline = tests/vcf_examples/1/1ped.xml
 coverage_report = tests/vcf_examples/1/1.coverage.pdf
+
 [gene_lists]
   [[MIT]]
     name = MIT
@@ -30,11 +31,11 @@ coverage_report = tests/vcf_examples/1/1.coverage.pdf
   [[OMIM-141201]]
     name = OMIM-141201
     version = 13.0
-#[individuals]
-#  [[1-1-3A]]
-#    name = 1-1-3A
-#    bam_path = /mnt/hds/proj/cust003/develop/analysis/exomes/1-1-3A/mosaik/gatk/1-1-3A.20100624_hiSeq20100624_X.lane1_sorted_pmd_rreal_brecal.bam
-#  [[1-1-4A]]
-#    name = 1-1-4A
-#    bam_path = /mnt/hds/proj/cust003/develop/analysis/exomes/1-1-4A/mosaik/gatk/1-1-4A.20100624_hiSeq20100624_X.lane1_sorted_pmd_rreal_brecal.bam
-#
+
+[individuals]
+ [[1-1-3A]]
+   name = 1-1-3A
+   bam_path = /mnt/hds/proj/cust003/develop/analysis/exomes/1-1-3A/mosaik/gatk/1-1-3A.20100624_hiSeq20100624_X.lane1_sorted_pmd_rreal_brecal.bam
+ [[1-1-4A]]
+   name = 1-1-4A
+   bam_path = /mnt/hds/proj/cust003/develop/analysis/exomes/1-1-4A/mosaik/gatk/1-1-4A.20100624_hiSeq20100624_X.lane1_sorted_pmd_rreal_brecal.bam
